+++ processor/src/test/fixtures/bad_input/com/example/injectable/QualifiedWithNoTypesModel.java
@@ -1,0 +1,14 @@
+package com.example.injectable;
+
+import sting.Eager;
+import sting.Injectable;
+import sting.Named;
+import sting.Typed;
+
+@Injectable
+@Named( "X" )
+@Typed( {} )
+@Eager
+class QualifiedWithNoTypesModel
+{
+}